Chore(): View 패키지 생성
View패키지를 생성하고 input, output클래스를 넣었습니다.
또한 Output 클래스의 역할이 미미해 Input으로 합치고 View로 이름을
변경했습니다.
